// Appraisal-Relevant-Information/dict/api/relevantData/types.ts
/* eslint-disable */
// tslint:disable
/**
 * Appraisal Service
 * Appraisal Service
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator+.
 * https://github.com/karlvr/openapi-generator-plus
 * Do not edit the class manually.
 */

import { Express } from 'express'
import { Api } from '../../models'

export interface RelevantDataApi {
	postRelevantDataCreate: (request: Api.RelevantDataDto | undefined) => Promise<PostRelevantDataCreateResponse>
	deleteRelevantDataDelete: (id: string) => Promise<DeleteRelevantDataDeleteResponse>
	getRelevantDataGet: (id: string) => Promise<GetRelevantDataGetResponse>
	getRelevantDataGetAll: () => Promise<GetRelevantDataGetAllResponse>
	putRelevantDataUpdate: (id: string, request: Api.RelevantDataDto | undefined) => Promise<PutRelevantDataUpdateResponse>
}

export type PostRelevantDataCreateResponse = PostRelevantDataCreate201Response | PostRelevantDataCreate401Response | PostRelevantDataCreate404Response | PostRelevantDataCreate422Response

export interface PostRelevantDataCreate201Response {
	status: 201
	body: Api.RelevantDataDto
	headers?: never
}

export interface PostRelevantDataCreate401Response {
	status: 401
	body: Api.MessageData
	headers?: never
}

export interface PostRelevantDataCreate404Response {
	status: 404
	body: Api.MessageData
	headers?: never
}

export interface PostRelevantDataCreate422Response {
	status: 422
	body: Api.MessageData
	headers?: never
}

export type DeleteRelevantDataDeleteResponse = DeleteRelevantDataDelete200Response | DeleteRelevantDataDelete401Response | DeleteRelevantDataDelete404Response | DeleteRelevantDataDelete422Response

export interface DeleteRelevantDataDelete200Response {
	status: 200
	body: Api.MessageData
	headers?: never
}

export interface DeleteRelevantDataDelete401Response {
	status: 401
	body: Api.MessageData
	headers?: never
}

export interface DeleteRelevantDataDelete404Response {
	status: 404
	body: Api.MessageData
	headers?: never
}

export interface DeleteRelevantDataDelete422Response {
	status: 422
	body: Api.MessageData
	headers?: never
}

export type GetRelevantDataGetResponse = GetRelevantDataGet200Response | GetRelevantDataGet401Response | GetRelevantDataGet404Response | GetRelevantDataGet422Response

export interface GetRelevantDataGet200Response {
	status: 200
	body: Api.RelevantDataDto
	headers?: never
}

export interface GetRelevantDataGet401Response {
	status: 401
	body: Api.MessageData
	headers?: never
}

export interface GetRelevantDataGet404Response {
	status: 404
	body: Api.MessageData
	headers?: never
}

export interface GetRelevantDataGet422Response {
	status: 422
	body: Api.MessageData
	headers?: never
}

export type GetRelevantDataGetAllResponse = GetRelevantDataGetAll200Response | GetRelevantDataGetAll401Response | GetRelevantDataGetAll404Response

export interface GetRelevantDataGetAll200Response {
	status: 200
	body: Api.RelevantDataPagedResultData
	headers?: never
}

export interface GetRelevantDataGetAll401Response {
	status: 401
	body: Api.MessageData
	headers?: never
}

export interface GetRelevantDataGetAll404Response {
	status: 404
	body: Api.MessageData
	headers?: never
}

export type PutRelevantDataUpdateResponse = PutRelevantDataUpdate200Response | PutRelevantDataUpdate401Response | PutRelevantDataUpdate404Response | PutRelevantDataUpdate422Response

export interface PutRelevantDataUpdate200Response {
	status: 200
	body: Api.RelevantDataDto
	headers?: never
}

export interface PutRelevantDataUpdate401Response {
	status: 401
	body: Api.MessageData
	headers?: never
}

export interface PutRelevantDataUpdate404Response {
	status: 404
	body: Api.MessageData
	headers?: never
}

export interface PutRelevantDataUpdate422Response {
	status: 422
	body: Api.MessageData
	headers?: never
}